blog雏形

+++ api/foo/api/blog_article.py
@@ -161,6 +161,27 @@ class BlogAritcleXHR(tornado.web.RequestHandler):
         logging.info(self.request)
         logging.info("got article_id %r in uri", article_id)
 
+        access_token = None
+        try:
+            access_token = self.request.headers['Authorization']
+            access_token = access_token.replace('Bearer ','')
+        except:
+            logging.info("got access_token null")
+            self.set_status(401) # Unauthorized
+            self.write('Unauthorized')
+            self.finish()
+            return
+        logging.info("got access_token %r", access_token)
+
+        token = auth_access_token_dao.auth_access_token_dao().query(access_token)
+        if token is None:
+            self.set_status(403) # Forbidden
+            self.write('Forbidden')
+            self.finish()
+            return
+
+        # TODO 检查文章是否account_id创建
+
         article = blog_article_dao.blog_article_dao().delete(article_id)
 
         self.set_status(200) # OK
@@ -225,7 +246,7 @@ class BlogAritcleXHR(tornado.web.RequestHandler):
         blog_article_dao.blog_article_dao().update(data)
 
         self.set_status(200) # OK
-        self.finish(JSON.dumps({'article_id':article_id}))
+        self.finish(JSON.dumps("OK"))
         return
 
 
@@ -256,4 +277,44 @@ class BlogAccountAritclesXHR(tornado.web.RequestHandler):
 
         self.set_status(200) # OK
         self.finish(JSON.dumps(articles))
+        return
+
+
+# /blog/articles/([a-z0-9]*)
+class BlogAritclePubXHR(tornado.web.RequestHandler):
+    # 发布文章
+    @tornado.web.asynchronous
+    @tornado.gen.coroutine
+    def put(self, article_id):
+        logging.info(self.request)
+        logging.info("got article_id %r in uri", article_id)
+
+        access_token = None
+        try:
+            access_token = self.request.headers['Authorization']
+            access_token = access_token.replace('Bearer ','')
+        except:
+            logging.info("got access_token null")
+            self.set_status(401) # Unauthorized
+            self.write('Unauthorized')
+            self.finish()
+            return
+        logging.info("got access_token %r", access_token)
+
+        token = auth_access_token_dao.auth_access_token_dao().query(access_token)
+        if token is None:
+            self.set_status(403) # Forbidden
+            self.write('Forbidden')
+            self.finish()
+            return
+
+        # TODO 检查文章是否account_id创建
+
+        timestamp = int(time.time());
+        data = {'_id':article_id, 'status':'pub',
+                'last_update_time':timestamp, 'publish_time':timestamp}
+        blog_article_dao.blog_article_dao().update(data)
+
+        self.set_status(200) # OK
+        self.finish(JSON.dumps("OK"))
         return
